+++ flang/runtime/CMakeLists.txt
@@ -37,7 +37,9 @@ add_flang_library(FortranRuntime PARTIAL_SOURCES_INTENDED
   edit-input.cpp
   edit-output.cpp
   environment.cpp
+  extrema.cpp
   file.cpp
+  findloc.cpp
   format.cpp
   internal-unit.cpp
   iostat.cpp
@@ -50,8 +52,10 @@ add_flang_library(FortranRuntime PARTIAL_SOURCES_INTENDED
   numeric.cpp
   random.cpp
   reduction.cpp
+  product.cpp
   stat.cpp
   stop.cpp
+  sum.cpp
   terminator.cpp
   tools.cpp
   transformational.cpp
